mac - 2025.10.23.5: changed to remove hex file on error, to avoid make build issues

+++ mac/macin.c
@@ -71,7 +71,7 @@ hash (register const char *str, register size_t len)
       187, 187, 187, 187, 187, 187, 187, 187, 187, 187,
       187, 187, 187, 187, 187, 187, 187, 187, 187, 187
     };
-  register unsigned int hval = len;
+  register unsigned int hval = (unsigned int)len;
 
   switch (hval)
     {
